+++ json_serializable/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.5.3
+
+* If `JsonKey.fromJson` function parameter is `Iterable` or `Map` with type
+   arguments of `dynamic` or `Object`, omit the arguments when generating a
+   cast. 
+   `_myHelper(json['key'] as Map)` instead of
+   `_myHelper(json['key'] as Map<dynamic, dynamic>)`.  
+
 ## 0.5.2
 
 * If `fromJson`/`toJson` are set in `JsonKey`, apply them before any custom